+++ packages/eyes-api/src/input/CheckSettings.ts
@@ -55,7 +55,7 @@ export type CheckSettings<TElement, TSelector> = {
   visualGridOptions?: {[key: string]: any}
   hooks?: {beforeCaptureScreenshot: string}
   renderId?: string
-  variantId?: string
+  variationGroupId?: string
   timeout?: number
 }
 
@@ -157,7 +157,7 @@ export class CheckSettingsFluent<TElement = unknown, TSelector = unknown> {
       Object.entries(settings.visualGridOptions).forEach(([key, value]) => this.visualGridOption(key, value))
     }
     if (settings.renderId) this.renderId(settings.renderId)
-    if (settings.variantId) this.variantId(settings.variantId)
+    if (settings.variationGroupId) this.variationGroupId(settings.variationGroupId)
     if (!utils.types.isNull(settings.timeout)) this.timeout(settings.timeout)
   }
 
@@ -494,9 +494,9 @@ export class CheckSettingsFluent<TElement = unknown, TSelector = unknown> {
     return this
   }
 
-  variantId(variantId: string): this {
-    utils.guard.isString(variantId, {name: 'variantId'})
-    this._settings.variantId = variantId
+  variationGroupId(variationGroupId: string): this {
+    utils.guard.isString(variationGroupId, {name: 'variationGroupId'})
+    this._settings.variationGroupId = variationGroupId
     return this
   }