+++ packages/eyes-playwright/src/spec-driver.ts
@@ -59,8 +59,8 @@ export async function isEqualElements(frame: Context, element1: Element, element
 
 // #region COMMANDS
 
-export async function executeScript(frame: Context, script: ((...args: any) => any) | string, arg: any): Promise<any> {
-  script = utils.types.isString(script) ? (new Function(script) as (...args: any) => any) : script
+export async function executeScript(frame: Context, script: ((arg: any) => any) | string, arg: any): Promise<any> {
+  script = utils.types.isString(script) ? (new Function(script) as (arg: any) => any) : script
   const result = await frame.evaluateHandle(script, arg)
   return handleToObject(result)
 }
@@ -118,13 +118,9 @@ export async function type(frame: Context, element: Element | Selector, keys: st
   if (isSelector(element)) element = await findElement(frame, element)
   await element.type(keys)
 }
-export async function hover(
-  frame: Context,
-  element: Element | Selector,
-  position?: {x: number; y: number},
-): Promise<void> {
+export async function hover(frame: Context, element: Element | Selector): Promise<void> {
   if (isSelector(element)) element = await findElement(frame, element)
-  await element.hover({position})
+  await element.hover()
 }
 export async function scrollIntoView(frame: Context, element: Element | Selector, align = false): Promise<void> {
   if (isSelector(element)) element = await findElement(frame, element)