+++ packages/eyes-testcafe/index.ts
@@ -1,5 +1,0 @@
-import {Eyes} from './src/api'
-
-export * from './src/api'
-
-export default Eyes