+++ packages/eyes-puppeteer/CHANGELOG.md
@@ -5,6 +5,16 @@
 - added full typescript support
 - introduced @applitools/eyes-api package with new api
 
+## 1.6.0 - 2021/4/26
+
+- updated to @applitools/eyes-sdk-core@12.17.4 (from 12.17.2)
+- updated to @applitools/visual-grid-client@15.8.2 (from 15.8.0)
+
+## 1.5.0 - 2021/4/22
+
+- updated to @applitools/eyes-sdk-core@12.17.2 (from 12.14.2)
+- updated to @applitools/visual-grid-client@15.8.0 (from 15.5.14)
+
 ## 1.4.3 - 2021/1/29
 
 - chore: add husky