+++ packages/sdk-coverage-tests/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 
+## 2.3.11 - 2021/4/26
+
+- include custom tests in report
+
 ## 2.3.9 - 2021/4/1
 
 - add `api` parameter to test object in bongo report