+++ packages/visual-grid-client/CHANGELOG.md
@@ -3,6 +3,19 @@
 ## Unreleased
 
 
+## 15.8.2 - 2021/4/26
+
+- rename `variantId` to `variationGroupId`
+- updated to @applitools/eyes-sdk-core@12.17.4 (from 12.17.2)
+
+## 15.8.1 - 2021/4/22
+
+- fix types for eyes open
+
+## 15.8.0 - 2021/4/22
+
+- added support for batch properties
+
 ## 15.7.1 - 2021/4/6
 
 - updated to @applitools/eyes-sdk-core@12.17.2 (from 12.17.0)